+++ app/src/androidTest/java/com/leevinapp/testingdemo/MainFragmentTest.kt
@@ -1,17 +1,22 @@
 package com.leevinapp.testingdemo
 
 import androidx.test.espresso.Espresso
-import androidx.test.espresso.assertion.ViewAssertions
+import androidx.test.espresso.assertion.ViewAssertions.matches
 import androidx.test.espresso.matcher.ViewMatchers
+import androidx.test.espresso.matcher.ViewMatchers.isCompletelyDisplayed
+import androidx.test.espresso.matcher.ViewMatchers.isDisplayed
 import androidx.test.espresso.matcher.ViewMatchers.withId
 import androidx.test.espresso.matcher.ViewMatchers.withText
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.rule.ActivityTestRule
-import okhttp3.mockwebserver.Dispatcher
 import com.leevinapp.testingdemo.utils.ResourceFile
+import com.leevinapp.testingdemo.utils.ViewIdlingResource
+import okhttp3.mockwebserver.Dispatcher
 import okhttp3.mockwebserver.MockResponse
 import okhttp3.mockwebserver.MockWebServer
 import okhttp3.mockwebserver.RecordedRequest
+import org.hamcrest.Matchers.allOf
+import org.hamcrest.Matchers.not
 import org.junit.After
 import org.junit.Before
 import org.junit.Rule
@@ -20,17 +25,18 @@ import org.junit.runner.RunWith
 import java.util.concurrent.TimeUnit
 
 @RunWith(AndroidJUnit4::class)
-class MainActivityTest {
+class MainFragmentTest {
 
     @get:Rule
-    val activityRule = ActivityTestRule(MainActivity::class.java, true, false)
+    val testActivityRule = ActivityTestRule(SingleFragmentActivity::class.java, true)
+    lateinit var fragment:MainFragment
+    @Before
+    fun setUp() {
+        mockWebServer.start(8000)
+        fragment = MainFragment()
+    }
 
     private val mockWebServer = MockWebServer()
-
-    @Before
-    fun setup() {
-        mockWebServer.start(8080)
-    }
 
     @After
     fun tearDown() {
@@ -42,18 +48,21 @@ class MainActivityTest {
         mockWebServer.dispatcher = object : Dispatcher() {
             override fun dispatch(request: RecordedRequest): MockResponse {
                 return MockResponse()
-                        .setResponseCode(200)
-                        .setBody(loadLocalResponse("success_response.json"))
+                    .setResponseCode(200)
+                    .setBody(loadLocalResponse("success_response.json"))
             }
         }
-        activityRule.launchActivity(null)
+
+        testActivityRule.activity.replaceFragment(fragment)
+
+        ViewIdlingResource.waitUntilIdleMatcherIsFulfilled(withId(R.id.textviewSuccess), isDisplayed())
 
         Espresso.onView(withId(R.id.progress_bar))
-                .check(ViewAssertions.matches(ViewMatchers.withEffectiveVisibility(ViewMatchers.Visibility.GONE)))
+            .check(matches(not(isDisplayed())))
         Espresso.onView(withId(R.id.textviewSuccess))
-                .check(ViewAssertions.matches(ViewMatchers.withEffectiveVisibility(ViewMatchers.Visibility.VISIBLE)))
+            .check(matches(isDisplayed()))
         Espresso.onView(withId(R.id.textview))
-                .check(ViewAssertions.matches(ViewMatchers.withEffectiveVisibility(ViewMatchers.Visibility.GONE)))
+            .check(matches(not(isDisplayed())))
     }
 
     @Test
@@ -64,16 +73,17 @@ class MainActivityTest {
             }
         }
 
-        activityRule.launchActivity(null)
+        testActivityRule.activity.replaceFragment(fragment)
+        ViewIdlingResource.waitUntilIdleMatcherIsFulfilled(allOf(withId(R.id.textview), withText(R.string.error_message)), isCompletelyDisplayed())
 
         Espresso.onView(withId(R.id.progress_bar))
-                .check(ViewAssertions.matches(ViewMatchers.withEffectiveVisibility(ViewMatchers.Visibility.GONE)))
+            .check(matches(not(isDisplayed())))
         Espresso.onView(withId(R.id.textviewSuccess))
-                .check(ViewAssertions.matches(ViewMatchers.withEffectiveVisibility(ViewMatchers.Visibility.GONE)))
+            .check(matches(not(isDisplayed())))
         Espresso.onView(withId(R.id.textview))
-                .check(ViewAssertions.matches(ViewMatchers.withEffectiveVisibility(ViewMatchers.Visibility.VISIBLE)))
+            .check(matches(isDisplayed()))
         Espresso.onView(withId(R.id.textview))
-                .check(ViewAssertions.matches(withText(R.string.error_message)))
+            .check(matches(withText(R.string.error_message)))
     }
 
     @Test
@@ -85,21 +95,20 @@ class MainActivityTest {
                     .setBody("[]")
             }
         }
-
-        activityRule.launchActivity(null)
+        testActivityRule.activity.replaceFragment(fragment)
+        ViewIdlingResource.waitUntilIdleMatcherIsFulfilled(allOf(withId(R.id.textview), withText(R.string.no_data_message)), isCompletelyDisplayed())
 
         Espresso.onView(withId(R.id.progress_bar))
-            .check(ViewAssertions.matches(ViewMatchers.withEffectiveVisibility(ViewMatchers.Visibility.GONE)))
+            .check(matches(ViewMatchers.withEffectiveVisibility(ViewMatchers.Visibility.GONE)))
         Espresso.onView(withId(R.id.textviewSuccess))
-            .check(ViewAssertions.matches(ViewMatchers.withEffectiveVisibility(ViewMatchers.Visibility.GONE)))
+            .check(matches(ViewMatchers.withEffectiveVisibility(ViewMatchers.Visibility.GONE)))
         Espresso.onView(withId(R.id.textview))
-            .check(ViewAssertions.matches(ViewMatchers.withEffectiveVisibility(ViewMatchers.Visibility.VISIBLE)))
+            .check(matches(ViewMatchers.withEffectiveVisibility(ViewMatchers.Visibility.VISIBLE)))
         Espresso.onView(withId(R.id.textview))
-            .check(ViewAssertions.matches(withText(R.string.no_data_message)))
+            .check(matches(withText(R.string.no_data_message)))
     }
 
     fun loadLocalResponse(path:String):String {
         return ResourceFile(path).readText()
     }
-
 }