+++ app/src/androidTest/java/com/leevinapp/testingdemo/TestApplication.kt
@@ -4,6 +4,6 @@ import android.app.Application
 
 class TestApplication : Application(), BaseUrlProvider {
     override fun getBaseUrl(): String {
-        return "http://localhost:8080"
+        return "http://localhost:8000"
     }
 }