+++ app/src/androidTest/java/com/leevinapp/testingdemo/utils/ResourceFile.kt
@@ -1,4 +1,4 @@
-package com.leevinapp.testingdemo
+package com.leevinapp.testingdemo.utils
 
 import java.nio.charset.Charset