+++ app/src/main/java/com/leevinapp/testingdemo/MainFragment.kt
@@ -1,27 +1,48 @@
 package com.leevinapp.testingdemo
 
-import android.content.Intent
 import android.os.Bundle
+import android.view.LayoutInflater
 import android.view.View
-import androidx.appcompat.app.AppCompatActivity
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import androidx.fragment.app.viewModels
+import androidx.test.espresso.idling.CountingIdlingResource
 import com.leevinapp.testingdemo.databinding.ActivityMainBinding
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
 
-class MainActivity : AppCompatActivity() {
+class MainFragment : Fragment() {
     private lateinit var binding: ActivityMainBinding
+    val countingIdlingResource = CountingIdlingResource("Main")
 
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        binding = ActivityMainBinding.inflate(layoutInflater)
-        setContentView(binding.root)
+    init {
+//        countingIdlingResource.increment()
+    }
 
-        val retrofit = RetrofitProvider(getBaseUrl(), OkHttpProvider.getOkHttpClient()).getRetrofit()
+    private val viewModel: MainViewModel by viewModels()
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+        binding = ActivityMainBinding.inflate(inflater, container, false)
+        return binding.root
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+        val retrofit =
+            RetrofitProvider(getBaseUrl(), OkHttpProvider.getOkHttpClient()).getRetrofit()
         val githubApi = retrofit.create(GithubApi::class.java)
 
+
         githubApi.testFetch().enqueue(object : Callback<List<GithubRepoResponse>> {
-            override fun onResponse(call: Call<List<GithubRepoResponse>>, response: Response<List<GithubRepoResponse>>) {
+            override fun onResponse(
+                call: Call<List<GithubRepoResponse>>,
+                response: Response<List<GithubRepoResponse>>
+            ) {
                 if (response.isSuccessful && response.body() != null) {
                     val result = response.body()!!
                     if (result.isEmpty()) {
@@ -38,10 +59,6 @@ class MainActivity : AppCompatActivity() {
                 showErrorState()
             }
         })
-
-        binding.openMainFragment.setOnClickListener {
-            startActivity(Intent(this,SingleFragmentActivity::class.java))
-        }
     }
 
     private fun showSuccessDataState(body: List<GithubRepoResponse>) {
@@ -66,6 +83,6 @@ class MainActivity : AppCompatActivity() {
     }
 
     private fun getBaseUrl(): String {
-        return (applicationContext as BaseUrlProvider).getBaseUrl()
+        return (activity?.applicationContext as BaseUrlProvider).getBaseUrl()
     }
 }