+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [x.x.x] - {Released / Unreleased}
 ### Added
 - Add UI testing with MockWebserver
+- Add 2 ways of fragment testing with MockWebserver
 ### Changed
 -
 ### Fixed